installer: query the Uninstaller values much earlier
We will want to access the existing config values in the next commit so
we better have a way to get at the existing Git for Windows location,
which is stored in the Uninstaller registry values.

Let's query those values already at the start, and store the values in
global variables.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>