installer: re-generate the system config instead of copying it
When Git for Windows is shipped as part of a bigger system (such as VFS
for Git), it can make a ton of sense to set system config variables. In
such a scenario, it is really, really unfortunate that Git for Windows'
installer simply blows it away and copies its own version over it during
an upgrade.

To help with that, let's change the logic. Let's teach the installer to
work on an existing system config (if any), setting the system config
values explicitly via `git config calls`.

Note: this does not yet save the current system config when upgrading
Git for Windows; That task is left for another commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>