installer: use GitSystemConfigSet() also to unset config vars
It is logically almost the same, and even the implementation is almost
the same. The only sadness is that we cannot use Nil values to indicate
that we want to unset the variable, but NUL is almost as good.

This ties up a loose end left by the earlier implementation of the
GitSystemConfigSet() function; We did not really need it, but it makes
things more readable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>