installer: retain system config across upgrades
In particular in complex setups like VFS for Git, where Git for Windows
is just one installed component, it does makes sense for 3rd-party
software to configure settings in Git for Windows' system config.

In such scenarios, those settings should be retained between Git for
Windows upgrades as much as possible.

So lets' do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>