please.sh upgrade: learn about git-flow
Git-flow v1.12.0 is out for a while, and MSYS2 even picked it up (after
we opened a PR...), but we still shadow MSYS2's version. And we should
keep doing so, as we have a bigger interest in git-flow being up to date
than MSYS2 does.

So let's make it more convenient to upgrade that package in Git for
Windows' SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>