ReleaseNotes: pull the update to msys2-runtime v3.0.1
v3.0.1-1 is broken in at least one, very fundamental way: the output of
`uname -m` misses the first character (and other fields in `uname -a`
seem to have the same problem).

This causes problems e.g. when `pacman` validates that a given package's
architecture matches the current system's.

At this moment, it is unclear where the problem was introduced, but this
close to Git for Windows v2.21.0-rc2, there is little else but to roll
back, and to leave it rolled back until after Git for Windows v2.21.0 is
released.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>