installer: use double quotes when configuring editors
The single quotes are fine when a *shell script* launches the editor,
but not when `git.exe` does.

This led to the funny issue where switching to the built-in as default
editor triggered this bug, because the built-in simply is not a shell
script and therefore cannot handle single quotes when spawning an
editor.

This fixes https://github.com/git-for-windows/git/issues/2011

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>