installer: detect (and prefer) system-wide Sublime Text installation
We support the Sublime Text editor for a while now, but only the
user-wide installation.

This developer was not even able to install Sublime Text in that way,
only the regular system-wide installation.

Sadly, there is no registry entry that we can detect, but the default
location is easy enough to figure out, and we can use it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>