git-extra: fix /usr/bin/start in the post-install script
The /usr/bin/start script relies on /usr/bin/cmd, or at least on cmd.exe
to be in the PATH. Due to the PATH containing %SystemRoot% entries in
many cases, this does not work (maybe it did work with a previous MSYS2
runtime?).

But Git for Windows does not include /usr/bin/cmd.

Let's just use $COMSPEC directly.

This fixes https://github.com/git-for-windows/git/issues/2014

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>