sdk: fix builds of PKGBUILD-managed projects
We do not really have any, yet, except for the `msys2-runtime`, and we
will change that one in the next commit to allow for building an
msys0.dll with local changes (i.e. not a full msys2-runtime Pacman
package), but we still should first cd to the correct directory, and we
still should return the exit code (this was a late-night typo).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>