sdk: allow for building msys-2.0.dll with local modifications
Actually, it builds msys0.dll ;-) But that one can be copied over to a
fresh SDK, or Portable Git, for testing.

Point being: if there are local modifications, then we should not build
the package (tossing away local modifications), but just the .dll, as
that is more likely what the user wants.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>