ReleaseNotes: do not mention Cygwin v3.x just yet
It is still too soon after Git for Windows v2.21.0 was released; we
might need to release a v2.21.0(2) soon, or a v2.21.1. And I am not
willing to risk the MSYS2 runtime upgrade just yet.

So the `/etc/pacman.conf` of Git for Windows' SDK will keep ignoring the
msys2-runtime upgrade for now, and there won't be any Git for Windows
version with such an MSYS2 runtime in, say, the next two weeks.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>