git askyesno: imitate gui--askyesno's command-line handling
We are trying to come up with a drop-in replacement for `git
gui--askyesno` because that command is not accessible (read: a screen
reader cannot make sense of it).

To make it a true drop-in replacement, we also need to imitate how `git
gui--askyesno` accepted multiple command-line parameters and simply
joined them into a single string.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>