git update-git-for-windows: use accessible popup
When we ask whether the user is ready to install an update, we use a
graphical popup. So far, we used `git gui--askyesno`, which was
convenient because it was small, and we ship with Tcl/Tk anyway.

However, that GUI helper is not nice to screen readers, making it
inaccessible.

We just introduced a drop-in replacement that is accessible, let's use
that one instead.

This fixes https://github.com/git-for-windows/git/issues/2099

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>