installer: fix `release.sh -d <page-title>` in a fresh clone
When no installer had been created before, there was no left-over
`package-versions.txt`, but we always tell the InnoSetup compiler to
bundle that file.

To work around that, let's generate an empty file of that name if we're
not asking `make-file-list.sh` to generate a real version of that file
(i.e. when we're skipping the files).

Noticed by Philip Oakley.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>