installer: fix version of installer and installed file
As reported by @JJClements, the File version and Product version of the
installer and installer binary were not the same. They are set by the
`installer/release.sh` and by using `sdk build (installer |
git-and-installer)` commands, which created this discrepancy.

Create two macros, `FILE_VERSION` and `PROD_VERSION` to store File and
Product version of the included binary and copy it to the installer
file.

Closes: git-for-windows/git#1797
Signed-off-by: Rohit Ashiwal <rohit.ashiwal265@gmail.com>