sdk: offer to edit a few well-known files
With this patch, you can call `sdk edit <file>` for a few files.

Well-known files are `git-sdk.sh` for the `sdk` function itself, and
`sdk.completion` for the tab completion in Bash for the `sdk` function.

More interestingly, users can now call

	sdk edit ReleaseNotes.md

to quickly open the release notes in the editor, without having to
remember where that file lives.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>