avoid Github API rate limit
The Github API has a per-user rate limit. For anonymous users, which we
are, the limit counts per IP address. For large organizations this can
be an issue. Instead, make the latest version check against
gitforwindows.org.

We still make calls to the Github API when we decide that we should
download a new release. But this happens only once per user and Git
release.

Signed-off-by: Clemens Buchacher <drizzd@gmx.net>