git-extra: offer an easy way to update an SDK via Pacman
Pacman can be relatively hard to use on Windows, as it was designed for
Linux, where files can be overwritten even if they are in use.

It gets even worse by having Pacman depend on the MSYS2 runtime: if that
runtime is overwritten while in use, all kinds of "interesting" issues
ensue, and the way MSYS2's Pacman handles it is not particularly
graceful (it tells the user to please kill Pacman and if it is running
inside a MinTTY, then that one, too, after upgrading core packages such
as `msys2-runtime`, `pacman` and `mintty`, when a better method might be
to move all of the overwritten files into the Recycle Bin first, so that
they do not actually have to be overwritten at all).

Besides, since MSYS2 tries to hardcode too much, Git for Windows has to
fix up a couple of things the hard way, via the post-install script that
runs as part of the `git-extra` installation/upgrade. So even after a
successful upgrade, it is necessary to re-install the `git-extra`
package.

Be that as it may, all of the above poses a number of challenges for
unsuspecting users of Git for Windows' SDK:

- The locally-synchronized package database might be out of date (Pacman
  looks at the mtime of the files, which may have been updated through
  a `git checkout` instead of by Pacman), and to fix that, the
  little-known double flag `-yy` needs to be passed to Pacman.

- Pacman might have to be run *twice* (when core packages need to be
  upgraded, it only updates those, and wants to be restarted to upgrade
  the rest).

- Not only Pacman, but also the Bash in which it was started, and the
  MinTTY window, too, potentially have to be closed (when upgrading
  "core" packages).

- After Pacman did its job, the `git-extra` package has to be updated,
  which is not obvious to most users.

To help with that, let's add a regular Batch script.

This makes it much easier to first terminate all MSYS2 processes that
would interfere with the upgrade (we might want to use the
blocked-file-util.exe helper in a future patch to make certain that all
those processes are closed) before starting the upgrade.

And putting all of that knowledge how to upgrade a Git for Windows SDK
into this script not only records that knowledge, but it does so in the
most useful form: as a fire-and-forget helper.

Existing users can copy the `update-via-pacman.bat` file directly into
their SDK's top-level directory, or get it via `pacman -Sy git-extra`
once https://github.com/git-for-windows/build-extra/pull/244 is merged.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>