Start implementing a meta credential helper
With this helper, configuring `credenial.helper = helper-selector` will
present the user with a list of all credential helpers found on the
`PATH` and then let the user choose which one to use.

It will also let the user persist their choice into the system config,
making it an ideal default for the portable flavor of Git for Windows.

This commit only implements a "Select" and a "Cancel" button, leaving
the reset to the next commits.

This will address https://github.com/git-for-windows/git/issues/2116

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>