sdk build: if run without an argument, try to guess on the basis of `pwd`
With this change, after `sdk build <project>` failed and the user fixed
a few things, a subsequent `sdk build` will not require the `<project>`
argument to be specified again.

Note: the logic is *very* simplistic at the moment, but seems to work
well enough to work in practice (it will not work in a subdirectory, but
then, the build will probably not work there, either).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>