sdk build: fix building e.g. `gnupg`
There are apparently a ton more variables that need to be adjusted...
see `/etc/msystem` and `/etc/profile` (in particular right after
`/etc/system` was sourced).

Let's replicate that here.

It would probably be the cleanest to move all of that
`MSYSTEM`-dependendt stuf from `/etc/profile` into `/etc/msystem`, but
that is outside the purview of this patch series.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>