ever-green: force cousins to be rebased
With frequent merging-rebases, there is one problem that is easy to
overlook: when merging branches that are not based on the "Start the
merging-rebase" commit, but instead on, say, a tag like v2.21.0, the
next merging-rebase needs to rebase this cousin. Otherwise, the `merge`
command will notice that this branch has *already* been merged (because
it was merged in the commit history that we made reachable by that fake
merge that started the merging-rebase).

Note: this commit will be overridden by the next commit, which does
things quite a bit differently. It is still here, for the historical
record, because a couple of branches have been handled by this version
of the ever-green.sh script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>