ever-green: avoid rebasing cousins again
There is a slightly more involved, and at the same time much more
accurate method: by using a replace ref to "replace away" the original
commit history.

The trick is to replace the fake merge that starts the merging-rebase by
an empty commit, temporarily, using `git replace --graft` just after
creating that merge commit, and deleting that graft just before the
rebase finishes.

Note: the sequencer is not yet clever enough to re-read replace objects
once they are cached. And this means that we have to forcefully exit the
rebase (following the same code path as if a `pick` failed because the
patch was already applied).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>