ever-green: avoid complicated shell expression
In dadd3215f (ever-green: make complicated shell expression more robust,
2019-04-17), we tried to make a complicated shell expression more robust
(it tries to add `^HEAD{/^Start the merging-rebase}` to the `rev-list`
invocation in a subshell, but only if we are in `--merging-rebase`
mode).

Turns out that it has not been made more robust: now it works in Git
Bash, but no longer in the Ubuntu agent in the Azure Pipeline that
should update the ever-green branches.

Let's just bite the bullet and introduce a nested subshell.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>