portable: avoid touching the Windows Credential Store by default
There are two major use cases for the portable flavor of Git for
Windows:

1. Users who want a user-wide installation, e.g. when they do not have
   administrator privileges on their machine, or when they want a
   specific version of Git for Windows for a specific task that does not
   override their installed version.

2. Users who want to install Git for Windows temporarily on somebody
   else's machine (or run it from a thumbdrive).

In the former case, our current default to use the Git Credential
Manager for Windows makes sense. In the latter case, not so much, as the
credentials will then be stored on the other users' machines.

To address this, we just integrated a meta credential helper, i.e. a
utility that discovers what credential helpers are available, lets the
user choose which one to use, and then use it.

Let's use it in the portable Git when available.

This addresses https://github.com/git-for-windows/git/issues/2116

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>