Mention New Feature in release notes
PortableGit [now comes with a meta credential
helper](https://github.com/git-for-windows/git/issues/2116), i.e. a
GUI that lets the user choose *which* of the available credential
helpers to use. This should help to avoid storing credentials on
other people's machines when running portable Git from a thumb drive.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>