Remove extra space between desc and plink edit box
This change reduces the space between the description text and the edit
box. The current install has a bit more space but adding an extra line
to the description and the Tortoise plink checkbox got a bit crowded.
Initially, the checkbox was partially visible but there was no scroll
bar.

Signed-off-by: Chris. Webster <chris@webstech.net>