installer/portable: add `/dev/` directory during setup
In b0c033bf3 (mkdir: always check-for-existence, 2019-06-03), Cygwin
alters the behavior of `mkdir` such that it will no longer create an
entry in the Windows filesystem for Cygwin-only directories like
`/dev/`. This commit was introduced in Git for Windows as part of the
upgrade to by moving to msys2-runtime v3.0.7 patch 3.

Previously, `mkdir` succeeded, allowing the post-install script
`01-devices.post` to both create `/dev` and successfully add files like
`/dev/fd` to it. Now, it fails to create the directory with the error

	mkdir: cannot create directory '/dev': File exists

(although this error is not seen because it pipes stderr to
`/dev/null`), and the script proceeds because `/dev` does effectively
exist and passes the existence check. However, by not being an actual
directory in the Windows filesystem, Cygwin doesn't allow creation of
any files in that directory, and returns a "Read-only file system"
error. Therefore creation of the `shm` and `mqueue` directories fail,
and adding the various `/dev/fd` links also fails. When this is called
by `post-install.bat`, the output is:

	"running post-install"
	mkdir: cannot create directory '/dev/shm': Read-only file system

	Creating /dev/shm directory failed.
	POSIX semaphores and POSIX shared memory will not work

	mkdir: cannot create directory '/dev/mqueue': Read-only file system

	Creating /dev/mqueue directory failed.
	POSIX message queues will not work

	ln: failed to create symbolic link '/dev/stdin': Read-only file system
	ln: failed to create symbolic link '/dev/stdout': Read-only file system
	ln: failed to create symbolic link '/dev/stderr': Read-only file system
	ln: failed to create symbolic link '/dev/fd': Read-only file system
	'C:\windows\system32\drivers\etc\hosts' -> '/etc/hosts'
	'C:\windows\system32\drivers\etc\protocol' -> '/etc/protocols'
	'C:\windows\system32\drivers\etc\services' -> '/etc/services'
	'C:\windows\system32\drivers\etc\networks' -> '/etc/networks'

Because this script has no error handling, the Git for Windows installer
thinks that `post-install.bat` succeeded when really it should have
failed.

As per https://cygwin.com/ml/cygwin/2019-08/msg00339.html, Cygwin
recommends that we create that directory as part of the installation, so
now we do.

Signed-off-by: Stephen Provine <stephpr@microsoft.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>