installer: also create /dev/{fd,stdin,stdout,stderr}
These symbolic links are expected by at least some MSYS2 programs, and
are therefore generated via the `01-devices.post` script.

However, rather than relying on the post-install script to create them,
we can do this ourselves.

Note that these are the symlinks emulated by Cygwin: system files that
start with the special prefix `!<symlink>\xff\xfe` in all-ASCII,
followed by the symlink target in UTF-16 format including the trailing
(UTF-16) NUL.

As 7-Zip and .zip files do not retain the system bit (at least not
reliably), we cannot do much about this in the portable Git and the
NuGet edition of Git for Windows.

This fixes https://github.com/git-for-windows/git/issues/2291

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>