git-extra: adjust location of the system config
In https://github.com/git-for-windows/git/pull/2358, we adjusted the
location of the system config and gitattributes file.

This patch adjusts the post-install script of the `git-extra` package
(which is used to edit some files that cannot otherwise be overridden in
a regular MSYS2 setup) to expect the system config in the new place.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>