installer: retain the system config across upgrades in the new location
The installer gained this neat feature where it retains any settings
across upgrades, even those that were configured in the system config by
something else than the Git for Windows installer.

We already look both in the old and the new location (which was the
location in Git for Windows 1.x) when saving a copy, but we now have to
use the location reported by Git when restoring it (depending which Git
version is about to be bundled in the installer, that is, which might be
the new one, `C:\Program Files\Git\etc\gitconfig`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>