installer: query the current `git.exe` for the system config location
In https://github.com/git-for-windows/git/pull/2358, we fixed the design
where Git for Windows would first read a "public" system config that was
not even in `C:\Program Files\Git` and then would read a "private"
system config. It now only reads _one_ system config, and that is in an
intuitive location: `C:\Program Files\Git\etc\gitconfig`.

Let's teach the script that generates the installer to make sure to ask
`git.exe` what its idea of the system config location is, and then use
that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>