mingit: include Git for Windows' system config
Previously, we shared Git for Windows' config with MinGit-based
applications via `C:\ProgramData\Git\config`.

However, we just reconciled this config with the "other" system config,
so there is no system config outside of the installation anymore.

To reinstate the ability of MinGit to reuse the settings configured via
Git for Windows' installer (such as `core.autoCRLF`), let's just include
the system config of the installed Git for Windows (if any) right at the
beginning of MinGit's own system config.

For good measure, we first include the system config of a 32-bit Git for
Windows that is installed on a 64-bit Windows (which does not make too
much sense, but some users still do that), and then include the regular
Git for Windows' system config.

In addition to hard-coding the `C:\Program Files` part, we also
hard-code the path to the new location of the system config (i.e. the
one _without_ the `mingw64` part) because the Git for Windows
installation will most likely have it long before any new MinGit version
meets the users.

Note: if `include.path` does not point to an existing file, Git will
simply ignore it, without a warning, which is exactly what we want here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>