installer: do not bother with `C:\ProgramData\Git\config` anymore
In https://github.com/git-for-windows/git/pull/2358, we reconciled the
two system configs into a single one.

Even if we now build an installer bundling a Git executable that is
still ProgramData-aware, it does not hurt to already drop bothering
about editing/initializing it: the system config in `C:\Program Files`
would override the ProgramData one anyway.

Note: for the time being, we still initialize the defaults for the
settings that can be configured in Git for Windows' installer from
_both_ ProgramData _and_ the previous installation's system config.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>