git-extra: ensure that /ssl is a symlink to /usr/ssl
With the recent change that moved Git for Windows' system gitconfig to
`/etc/gitconfig`, we now share that file with the MSYS variant of
`git.exe`.

That gitconfig contains the absolute path of the OpenSSL certificates
when fetching via HTTPS (and when `http.sslBackend` says that we should
use OpenSSL), or at least something that _looks_ like an absolute _Unix_
path: it is actually interpreted as relative to the runtime prefix.

The path in question is: `/ssl/certs/ca-bundle.crt`.

The MSYS variant of `git.exe`, however, will interpret this relative to
the MSYS2 root, but that's not where that `ssl/` directory lives in
MSYS2: it lives in `/usr/ssl/` instead.

And those certificates matter: the MSYS `git.exe` is compiled only with
the OpenSSL backend (not with the Secure Channel one).

Let's just make sure to install a symlink so that even the MSYS
`git.exe` finds those certificates.

Note: we use a symlink, not a junction, as Git still has some problems
handling junctions properly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>