installer: do not warn about a "downgrade" from -rc<N> versions
When upgrading from, say, v2.24.0-rc2 to v2.24.0, we definitely do _not_
want a bogus "do you want to downgrade" warning.

To help the existing test, let's include the output of the `git version`
command and compare it with the just-enhanced `VersionCompare()`
function to the output of the to-be-upgraded Git for Windows version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>