installer: do detect downgrades _to_ -rc<N> versions
The `-rc<N>` versions reduce to a registry value `CurrentVersion` that
looks like it is a _later_ version than the final version. Example:
v2.24.0.windows.1 would reduce to `2.24.0.1` while
v2.24.0.rc2.windows.1` would reduce to `2.24.0.2.1`, i.e. the -rc2
version would be mistaken for being an _upgrade_ relative to the final
v2.24.0.

To fix this, let's detect whether the version we are about to install,
or the version we want to upgrade, are potentially `-rc<N>` versions,
simply by counting the dots, and fall back to the (a little bit more
expensive) test via `git version` if that is the case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>