Use the self-extract 7zip package instead of using innounp
That aside, this commit contains a lot of changes that can make it
possible to successfully get or update git-on-windows on `Msys2` or
`Cygwin` which is generally tested. Also, a feature of generating
scripts of adding and removing context menu items (`Git Bash Here` and
`Git GUI Here` is included in this commit so that user could easily add
or remove these two context menu items as they wish. Last but not least,
the two scripts are now one single file which makes it possible to call
this script without downloading via a single command `curl ... | bash`

Signed-off-by: liuxy <391861737@qq.com>