Ensure the `/cmd` folder exists and `sh` could be found by `Git GUI Here`
So that the generating feature won't fail when there's now /cmd folder
available, and fixes the `hook execution requires sh (not in PATH)`
error when using `Git GUI Here` on Cygwin whose default bin folder is
`/bin` rather than `/usr/bin`.

In addition:
There's a known issue that are hard to fix without breaking the
principle of `keeping every existing things intact`: the `Git Bash Here`
on Cygwin won't have a nice prompt which shows the current branch of
current directory. To fix this, the `/etc/bash.bashrc` needs be edited.

Signed-off-by: liuxy <391861737@qq.com>