mingit: fix system gitconfig layering
The original idea of the now-discontinued support for
`C:\ProgramData\Git\config` was to have a central place where Git for
Windows writes the settings configured via its installer, so that e.g.
MinGit can pick them up.

This is particularly important when the user changed the default of e.g.
`core.autocrlf` because running `git status` with the Git Bash and with
MinGit would otherwise produce inconsistent results.

To continue with this idea, MinGit now adds two `include.path` settings
that pick up the system gitconfig of the 32-bit and the 64-bit Git for
Windows, respectively, if any exist.

There is just one problem with that: we include those system gitconfigs
_first_, and then configure things like `core.autocrlf`... which
overrides Git for Windows' settings.

Let's instead include Git for Windows' system gitconfig _after_ setting
our defaults.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>