git-extra: unset http.sslCAInfo in the system gitconfig
The previous commit stopped initializing the system gitconfig with that
setting that we no longer need.

With this patch, we now actively unset this setting in existing
gitconfigs.

Note: we are extra careful _not_ to use `git config` to do the editing
so that we allow for users who absolutely definitely want to have that
setting (they just need to add an extra space or something like that).

While at it, we also remove the hack where we installed a symlink
`/ssl/` -> `/usr/ssl/` to allow for sharing that setting between the
MSYS and the MINGW version of `git.exe`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>