prerelease: work around /usr/bin/git vs /mingw64/bin/git problem
A recent `makepkg` update uses a good trick to save space: after cloning
the Git sources into a bare clone in the same directory as the
`PKGBUILD` file, the worktree in the `src/` subdirectory uses the bare
repository as a reference.

In `mingw-w64-git`'s case, that means that an alternate is installed
into `src/git/` that points to the absolute path of the bare
repository's object directory.

However, `makepkg` insists on always using `/usr/bin/git.exe`, and
therefore that absolute path is an _MSYS_ one. When accessing that
`src/git/` worktree with the MINGW version of `git.exe`, that causes a
problem (as it would expect an absolute _Windows_ path). Symptom:

error: object directory /usr/src/MINGW-packages/mingw-w64-git/git/objects does not exist; check .git/objects/info/alternates

Let's work around that problem in `prerelease` by dissociating the
worktree from the alternate.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>