Installer: extend `ReplayChoice()` to detect new options
This patch introduces the concept of a "page set": a string consisting
of comma-delimited page IDs.

Then, two such sets are built up: a set containing all the pages with
custom options that we potentially want to show to the user, and a set
with all the pages containing previously-unseen options.

So far, this does not help anything yet, the next commit will introduce
a way to be able to skip pages that contain only options that the user
saw before and for which they already chose the desired option.

Signed-off-by: Denise Draper <draperd@acm.org>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>