please.sh create-sdk-artifact: optionally bundle everything
"Everything" is of course large, hence we compress it with `xz`, with
the maximal compression, to save both on space as well as on download
time.

The intention of this artifact is to back (still to be introduced) Azure
Pipelines that the Git for Windows project will use to build and publish
Pacman packages of components like cURL, OpenSSL, Git LFS, etc

The big advantage of using the `.tar.xz` file format is that it can be
easily extracted via `tar xJf` in an installed Git for Windows (which
Azure Pipelines offers on its hosted Windows agents to run Bash tasks).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>