pacman-helper: add a quick mode
Currently, pacman-helper is a bit clunky to use, as it insists on having
a local cache of all the current packages and signatures.

That is overkill: we typically have a very specific set of artifacts
built from one particular `PKGBUILD`.

To help that, let's introduce a command that downloads just the Pacman
dbs that are _actually_ needed, update them, then upload just the files
and the updated dbs.

This will be the work horse of the "Pacman package publishing" part of
the Azure Pipeline "Build and publish MINGW Pacman package" at
https://dev.azure.com/git-for-windows/git/_build?definitionId=32

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>