please.sh create-sdk-artifact: don't create incremental artifacts
Turns out that it is quite cumbersome to consume the incremental
artifacts. The idea was to save on space, but given that the build
artifacts are deduplicated, there is no much space saving in that.

So let's just go full artifacts: the `build-installers` artifact, for
example, will be enough to build `mingw-w64-git` and bundle installer,
PortableGit, MinGit and MinGit (BusyBox). (It won't be necessary any
longer to overlay it with the minimal-sdk, makepkg-git and
makepkg-git-i686 artifact.)

While at it, fold the `makepkg-git-i686` part into `makepkg-git`. We
pretty much always want to be able to build 64-bit *and* 32-bit versions
of `mingw-w64-git`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>