please.sh: avoid double slashes when expanding $pkgpath
First, we make sure that it always starts with a slash. This was already
the case except in a handful places where `usr/src/build-extra` was
passed to `up_to_date`.

Then, we remove the slash in all instances of `/$pkgpath`.

This helps the case where e.g. `$sdk32` is empty (which would result in
path like `//usr/src/build-extra`, which will be misinterpreted as a
network share).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>