please.sh upgrade: do not require GPG keys unless building
When building the packages, we may want to ensure that the local setup
knows a given GPG key known to be required to validate the source
packages.

But this is moot when we are not even building, which is the case in the
step in the Azure Pipeline that updates the `PKGBUILD` files and
friends. The build will only be performed later. And then we _will_ have
that key.

The code tries to be helpful and wants to import the missing GPG key and
lsign it. But that lsigning requires a GPG key of our own, which does
not exist in the Azure Pipeline (neither should it).

So let's not complain about those "missing" GPG keys when we ain't gonna
need 'em.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>