Installer: detect Developer Mode for symlink support detection
When running the installer from an Administrator account, currently the
detection fails whether the user can create symbolic links.

We now detect if the Developer Mode is enabled on Windows 10 Creators
Update or newer. If it is, we will check the box by default, both for
Admin and non-admin users. If Developer Mode is not enabled, we will
fall back to the old method.

Signed-off-by: Kristof Mattei <github@kristofmattei.be>