Enforce all text files to be checked out with LF-only line endings
Due to a switch to a redesigned Azure Pipeline that builds and publishes
our MINGW packages, the `git-extra` package is now built from a checkout
that uses the default `core.autoCRLF` setting of Git for Windows.

As a consequence, all files that were not marked explicitly with LF-only
line endings were changed to CR/LF line endings in the latest
`git-extra` package, including `git.nanorc` (and Nano does not like that
change, not one bit).

Let's just simplify the entire setup and tell Git that all text files
within `build-extra` are to have LF line endings.

This addresses https://github.com/git-for-windows/git/issues/2429

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>