Fix all text files that were committed with CR/LF line endings
We want all text files within `build-extra` to be checked out with
LF-only line endings, to avoid problems similar to the nanorc problem we
just fixed.

Originally, the plan was to keep the line endings of the `.isl` files,
but it seems that some of them have LF-only line endings and They Work
Just Fine, so let's not worry too much about keeping them with DOS-style
line endings.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>