create-shortcut: learn about Desktop shortcuts
It is actually not trivial for scripts to figure out the _actual_
location of the Desktop (it can be different from
`%USERPROFILE%\Desktop`, e.g. when letting OneDrive manage your
Desktop).

So let's introduce a new option for `create-shortcut.exe` so that we
have the full power of the Win32 API to figure this out:

	create-shortcut --desktop-shortcut [...]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>