installer: simplify the logic
When determining which page is the last one before the `Processes` page,
we are hard-coding the choice between the experimental options page (if
there are any experimental options) or the extra options page.

However, this hard-coding is totally unnecessary, as we do keep track of
the latest page we created, therefore it is just a matter of assigning
the variable at the exact right time.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>