installer: avoid expanding the `app` constant over and over again
In `RefreshProcessList()`, we want to make sure that a selected range of
`.exe` and `.dll` files are not currently in use because we are about to
overwrite them.

In this code, we expand the `app` constant over and over.

That turns out to be totally unnecessary, as we already have the global
variable `AppDir` ready for use.

A further advantage of using `AppDir` is that it is available even in
the `wpInfoBefore` page (i.e. the page where we show the GPL before even
starting the installer). This is in contrast to the `app` constant,
which is only available after the `wpSelectDir` page (whether that page
was shown or skipped).

As a consequence, we will be able to call `RefreshProcessList()` even on
that first wizard page, to determine whether we can skip essentially all
the pages because the user has seen all of the options before already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>