Drop old OpenSSL and LibIDN2 files
To enable a smooth transition from OpenSSL v1.0.x to v1.1.x, we kept the
old OpenSSL files around.

They are not in use anymore (except maybe in the Git for Windows SDK
when using `rhash.exe` via CMake, or when one tries to use the Qt5
components).

Likewise, we kept an older LibIDN2 version around (because it is needed
by GNU TLS which is used by cURL).

So let's stop retaining those old files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>