Force the Pacman packages back to use xz
A recent change in the MSYS2 project started to follow upstream Pacman
in using zstd to compress Pacman packages instead of xz:
https://github.com/msys2/MSYS2-packages/pull/1824

However, this is not yet a complete move: only MINGW packages are
switched, MSYS packages still use xz.

Git for Windows' tooling is pretty hard-coded in using the `.pkg.tar.xz`
extension. It would be relatively hard to transition to `.pkg.tar.zst`
at the moment, _in particular_ if the MSYS packages have a different
extension than the MINGW ones.

So let's override MSYS2's switch in Git for Windows' SDK, at least for
the moment.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>