installer: handle Vim 8.2
The installer bash script comments out a part of Vim's defaults.vim
file. This part resets the cursor line to the last one when reopening a
file. This is not desired in the case of commit messages, which are
always the same file name but composed from scratch each time.

In the meantime, MSYS2 Vim was updated to 8.2.

This causes a warning on installing git-extra because the installer
script unconditionally greps a Vim 8.1 file that is no longer present.
Fix this by checking the file's presence.

A similar installer script section is not necessary for Vim 8.2, because
this version includes in this line-resetting automatism an exception for
all file types matching "*commit*".